update to `NPC.py` file
makes some changes to `NPC.py` file, makes a master class that is meant to be used by all types of classes from the `NPC.py` file. This includes setting a consistent class function to get the text associated with each class, having a simple `self.open()` function as well as hopefully moving away from the previous system of making a new class each time a button is used, to passing the parent class object to buttons to reuse the same class over and over.